Switched to gaia-header
Changed the header from a custom header to gaia-toolbar. Whether or not
this was a good idea I don't really know, as the title in Gaia toolbar
is shifted to the left when there is an icon to the right side of it.
The title is centred on the connection screen, yet not centred on the
chat screen due to the close button.

extra:
Pressing connect on the connection screen will now bring you to the chat
screen. Pressing the close/disconnect button on the chat screen will
bring you back to the connection screen.